replace ckan with new catalog
